feat: fastify server
